refactor: tests and move into slow_tests top directory
