+++ slow_tests/README.md
@@ -1,0 +1,4 @@
+This directory contains long-running tests that should not
+be packaged with optlang. `data/miplib2003` contains all
+benchmark problems that can be solved <1 h with commercial
+solvers.
