--- tests/data/miplib2003.txt
@@ -1,18 +1,0 @@
-=opt=  10teams                         924
-=opt=  aflow30a                        1158
-=opt=  air05                           26374
-=opt=  arki001                         7580810
-=opt=  danoint                         65.67
-=opt=  disctom                         -5000
-=opt=  fiber                           405935.18000
-=opt=  fixnet6                         3983
-=opt=  gesa2                           25779856.372
-=opt=  gesa2-o                         25779856.372
-=opt=  glass4                          1.20001e+09
-=opt=  harp2                           -7.38998e+07
-=opt=  markshare1                      1
-=opt=  mas74                           11801.2
-=opt=  mas76                           40005.1
-=opt=  misc07                          2810
-=opt=  modglob                         2.07405e+07
-=opt=  momentum2                       12314.2
